Work on getting correct normal vector of a polygon
Having trouble getting a correct normal vector for a polygon. When I test the normal vector that is generated, it seems to be correct, but the cube is animating wrong... I think that I need to take the edges of the polygon to be tested as vectors and get the normal vector of those. Right now I'm using the points of the polygon...Maybe this is a problem of confusing points with vectors...